+++ pywm/pywm_widget.py
@@ -12,8 +12,8 @@ else:
 
 
 class PyWMWidgetDownstreamState:
-    def __init__(self, z_index: int=0, box: tuple[float, float, float, float]=(0, 0, 0, 0), mask: tuple[float, float, float, float]=(-1, -1, -1, -1), opacity: float=1., lock_enabled: bool=True, workspace: Optional[tuple[float, float, float, float]]=None, primitive: Optional[str]=None) -> None:
-        self.z_index = int(z_index)
+    def __init__(self, z_index: float=0, box: tuple[float, float, float, float]=(0, 0, 0, 0), mask: tuple[float, float, float, float]=(-1, -1, -1, -1), opacity: float=1., lock_enabled: bool=True, workspace: Optional[tuple[float, float, float, float]]=None, primitive: Optional[str]=None) -> None:
+        self.z_index = float(z_index)
         self.box = (float(box[0]), float(box[1]), float(box[2]), float(box[3]))
         self.mask = (float(mask[0]), float(mask[1]), float(mask[2]), float(mask[3]))
         self.opacity = opacity
@@ -24,7 +24,7 @@ class PyWMWidgetDownstreamState:
     def copy(self) -> PyWMWidgetDownstreamState:
         return PyWMWidgetDownstreamState(self.z_index, self.box, self.mask, self.opacity, self.lock_enabled, self.workspace)
 
-    def get(self, root: PyWM[ViewT], output: Optional[PyWMOutput], pixels: Optional[tuple[int, int, int, bytes]], primitive: Optional[tuple[str, list[int], list[float]]]) -> tuple[bool, tuple[float, float, float, float], tuple[float, float, float, float], int, float, int, tuple[float, float, float, float], Optional[tuple[int, int, int, bytes]], Optional[tuple[str, list[int], list[float]]]]:
+    def get(self, root: PyWM[ViewT], output: Optional[PyWMOutput], pixels: Optional[tuple[int, int, int, bytes]], primitive: Optional[tuple[str, list[int], list[float]]]) -> tuple[bool, tuple[float, float, float, float], tuple[float, float, float, float], int, float, float, tuple[float, float, float, float], Optional[tuple[int, int, int, bytes]], Optional[tuple[str, list[int], list[float]]]]:
         return (
             self.lock_enabled,
             root.round(*self.box, wh_logical=False),
@@ -54,7 +54,7 @@ class PyWMWidget(Generic[PyWMT]):
 
         self._pending_primitive: Optional[tuple[str, list[int], list[float]]] = None
 
-    def _update(self) -> tuple[bool, tuple[float, float, float, float], tuple[float, float, float, float], int, float, int, tuple[float, float, float, float], Optional[tuple[int, int, int, bytes]], Optional[tuple[str, list[int], list[float]]]]:
+    def _update(self) -> tuple[bool, tuple[float, float, float, float], tuple[float, float, float, float], int, float, float, tuple[float, float, float, float], Optional[tuple[int, int, int, bytes]], Optional[tuple[str, list[int], list[float]]]]:
         if self._damaged:
             self._damaged = False
             self._down_state = self.process()
